Update from original (#1)
* Added overview for each module

Because it was confusing to me which topics were discussed in week 1, week 2, etc. for Javascript 1 module. So I did the same for all the other modules. 

It might be duplicate in some parts, but maybe it's nice for clearing things up through the overview.

* Fixed small mistake with the tables

* I removed 'private', because apparently it is not

I, as an unauthorized user, can see the repo's readme, and this is the only file in the repo

* Add files via upload

First version of the Student Manual

* Update i-am-a-new-student.md

* Update i-am-a-new-student.md

* Delete Student Manual_v01_13052019.docx

* Student Manual V01 July 2 2019

* Delete Student Manual_v01_02072019.docx

* Update URL student manual and code of conduct

* Remove core values 

Soon to be updated with core values from Student Manual

* updated html/css + js1 section

* Updates links HTML/CSS & JS1

* added new descriptions and images to readme

* fixed hyperlinks to modules

* Update README.md

* Update README.md

Fix links to lesson plans for javascript 3

* added postgrad ed

* Update README.md

* Update README

Change student manual link

* added english booster + qa as addons

* update link week 0

* Update README.md

fixed the lessonplan links for javascript 2

Co-authored-by: marcianoviereck <marciano.viereck@simacan.com>
Co-authored-by: Igor Budasov <igor.budasov@gmail.com>
Co-authored-by: Wouter Kleijn <37488847+wouterkleijn@users.noreply.github.com>
Co-authored-by: Noer Paanakker <noer@hackyourfuture.net>
Co-authored-by: Noer <noerpaanakker@hotmail.com>
Co-authored-by: Tjebbee <33282467+Tjebbee@users.noreply.github.com>
Co-authored-by: Noer Paanakker <noer@lightworksweb.nl>